Crate timely_sort

Types and traits for radix sorting.

The current offerings are

LSBRadixSorter: A least-significant byte radix sorter.

MSBRadixSorter: A most-significant byte radix sorter.

LSBSWCRadixSorter: A least-significant byte radix sorter with software write combining.

There should probably be a MSBSWCRadixSorter in the future, because I like both of those things.

Structs

LSBRadixSorter
A few buffers capable of radix sorting by least significant byte.

LSBSWCRadixSorter
A few buffers capable of radix sorting by least significant byte.

MSBRadixSorter
A "most-significant byte" (MSB) radix sorter.

Traits

RadixSorter
Functionality provided by a radix sorter.

RadixSorterBase
Functionality independent of the type U used to sort.

Unsigned
An unsigned integer fit for use as a radix key.