Struct timely_sort::LSBRadixSorter

pub struct LSBRadixSorter<T> { /* fields omitted */ }

A few buffers capable of radix sorting by least significant byte.

The sorter allows the use of multiple different key bytes, determined by a type U: Unsigned. Currently, one is allowed to mix and match these as records are pushed, which may be a design bug.