Trait timely_sort::RadixSorterBase

pub trait RadixSorterBase<T> {
    fn new() -> Self;

    fn rebalance(&mut self, buffers: &mut Vec<Vec<T>>, intended: usize);

    fn recycle(&mut self, buffers: &mut Vec<Vec<T>>) { ... }
}

Functionality independent of the type U used to sort.

Required Methods

fn new() -> Self

Allocates a new instance of the radix sorter.

fn rebalance(&mut self, buffers: &mut Vec<Vec<T>>, intended: usize)

Provides empty buffers for the radix sorter to use, with the intent that it should own at most intended.

Provided Methods

fn recycle(&mut self, buffers: &mut Vec<Vec<T>>)

Provides empty buffers for the radix sorter to use.