Struct timely_sort::MSBRadixSorter

pub struct MSBRadixSorter<T> { /* fields omitted */ }

A "most-significant byte" (MSB) radix sorter.

This type manages buffers and provides logic for sorting batches of records using binary keys, in a "top down" or "most-significant byte" manner. One advantage of this approach is that it can exit early if batch sizes become small, avoiding radix scans over the data.

The type also allows the user to hook this the early exit functionality, as the finalizing action before the early exit can be a great time to consolidate or deduplicate records if this is desired, and can be much more efficient than trying to do the same after the fact (often the radix key is not the true key, and one must re-sort by the actual key). See the finish_into_and and sort_and methods.

Methods

impl<T> Sorter<T>

pub fn finish_into_and<U: Unsigned, F: Fn(&T) -> U, L: Fn(&mut Vec<T>)>(
&mut self,
target: &mut Vec<Vec<T>>,
bytes: F,
action: L
)

Finishes the sorting for the session, using the supplied finalizing action when given the option to exit early.

pub fn sort_and<U: Unsigned, F: Fn(&T) -> U, L: Fn(&mut Vec<T>)>(
&mut self,
source: &mut Vec<Vec<T>>,
bytes: F,
action: L
)

Radix sorts a sequence of buffers, possibly stopping early on small batches and calling action.

The intent of sort_and is to allow us to do top-down radix sorting with the option to exit early if the sizes of elements to sort are small enough. Of course, just exiting early doesn't sort the data, so we have an action you get to apply to finish things off. We could make this sort by radix, but there are several use cases where we need to follow up with additional sorting / compaction and would like to hook this clean-up method anyhow.